Move read-only patient work views to shared location
