+++ src/js/views/patients/shared/patient-readonly.scss
@@ -1,0 +1,26 @@
+
+.patient-readonly {
+  border: solid 1px transparent;
+  cursor: not-allowed;
+  display: inline-block;
+  font-size: 12px;
+  line-height: 14px;
+  overflow: hidden;
+  padding: 8px;
+  position: relative;
+  text-overflow: ellipsis;
+  vertical-align: middle;
+  white-space: nowrap;
+}
+
+.patient-readonly__status {
+  padding: 0 8px;
+}
+
+.patient-readonly__owner {
+  width: 80px;
+}
+
+.patient-readonly__time {
+  border-left: 1px solid $black-90;
+}
