+++ src/js/views/patients/worklist/action_views.js
@@ -5,6 +5,7 @@ import { View } from 'marionette';
 import 'scss/modules/table-list.scss';
 
 import { CheckComponent, StateComponent, OwnerComponent, DueComponent, TimeComponent, FormButton, DetailsTooltip } from 'js/views/patients/shared/actions_views';
+import { ReadOnlyStateView, ReadOnlyOwnerView, ReadOnlyDueDateView, ReadOnlyDueTimeView } from 'js/views/patients/shared/read-only_views';
 
 import ActionItemTemplate from './action-item.hbs';
 
@@ -19,51 +20,6 @@ const ActionEmptyView = View.extend({
       <h2>{{ @intl.patients.worklist.actionViews.actionEmptyView }}</h2>
     </td>
   `,
-});
-
-const ReadOnlyStateView = View.extend({
-  tagName: 'span',
-  className: 'worklist-list__readonly worklist-list__status',
-  template: hbs`<span class="action--{{ stateOptions.color }}">{{fa stateOptions.iconType stateOptions.icon}}</span>{{~ remove_whitespace ~}}`,
-  templateContext() {
-    const stateOptions = this.model.getState().get('options');
-
-    return {
-      stateOptions,
-    };
-  },
-});
-
-const ReadOnlyOwnerView = View.extend({
-  tagName: 'span',
-  className: 'worklist-list__readonly worklist-list__owner',
-  template: hbs`{{far "circle-user" classes="u-margin--r-8"}}<span>{{ owner }}</span>`,
-  templateContext() {
-    return {
-      owner: this.model.getOwner().get('name'),
-    };
-  },
-});
-
-const ReadOnlyDueDateView = View.extend({
-  tagName: 'span',
-  className: 'worklist-list__readonly',
-  template: hbs`
-    <span{{#if isOverdue}} class="is-overdue"{{/if}}>
-      {{far "calendar-days"}}{{#if due_date}}<span class="u-margin--l-8">{{formatDateTime due_date "SHORT" inputFormat="YYYY-MM-DD"}}</span>{{/if}}
-    </span>
-  `,
-  templateContext() {
-    return {
-      isOverdue: this.model.isOverdue(),
-    };
-  },
-});
-
-const ReadOnlyDueTimeView = View.extend({
-  tagName: 'span',
-  className: 'worklist-list__readonly worklist-list__time',
-  template: hbs`{{far "clock"}}{{#if due_time}}<span class="u-margin--l-8">{{formatDateTime due_time "LT" inputFormat="HH:mm:ss"}}</span>{{/if}}`,
 });
 
 const ActionItemView = View.extend({
